WarehouseReceiptDetail
edit service,dto,controller,entity

+++ app/backend/src/main/java/com/openedsource/pizzastore/controller/internal/WarehouseReceiptDetailController.java
@@ -15,7 +15,7 @@ import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
 
 @RestController
-@RequestMapping("/internal")
+@RequestMapping("/internal/warehousereceiptdetail")
 @CrossOrigin(origins = "*")
 public class WarehouseReceiptDetailController {
     @Autowired
@@ -23,7 +23,7 @@ public class WarehouseReceiptDetailController {
     @Autowired
     private WarehouseReceiptDetailService warehouseReceiptDetailService;
 
-    @PostMapping("/warehousereceiptdetail/add")
+    @PostMapping
     public ResponseEntity<Object> addWarehouseReceiptDetail(@RequestBody(required = false) WarehouseReceiptDetailDto warehouseReceiptDetailDto) {
         ResponseEntity<Object> response = ResponseEntity.status(HttpStatus.CREATED).build();
         String message = inputcheck(warehouseReceiptDetailDto);
@@ -33,6 +33,8 @@ public class WarehouseReceiptDetailController {
 
             try {
                 warehouseReceiptDetailService.insertWarehouseReceiptDetail(orderDetailEntity);
+                response = ResponseUtils.buildMessageReponse(HttpStatus.CREATED, "Successfully Added");
+
             } catch (DuplicateKeyException e) {
                 response = ResponseUtils.buildMessageReponse(HttpStatus.CONFLICT, Constants.MessageString.CONFLICT_ERROR.getMessage());
             }
@@ -42,15 +44,20 @@ public class WarehouseReceiptDetailController {
         return response;
     }
 
-    @PutMapping("/warehousereceiptdetail/update")
+    @PutMapping
     public ResponseEntity<Object> updateWarehouseReceiptDetail(@RequestBody(required = false) WarehouseReceiptDetailDto warehouseReceiptDetailDto) {
         ResponseEntity<Object> response = ResponseEntity.status(HttpStatus.OK).build();
         String message = inputcheck(warehouseReceiptDetailDto);
         if (ValidateUtils.isNullOrEmpty(message)) {
-            WarehouseReceiptDetailEntity orderDetailEntity = new WarehouseReceiptDetailEntity();
-            BeanUtils.copyProperties(warehouseReceiptDetailDto, orderDetailEntity);
-
-            warehouseReceiptDetailService.updateWarehouseReceiptDetail(orderDetailEntity);
+            WarehouseReceiptDetailEntity warehouseReceiptDetailEntity = new WarehouseReceiptDetailEntity();
+            BeanUtils.copyProperties(warehouseReceiptDetailDto, warehouseReceiptDetailEntity);
+            if (warehouseReceiptDetailRepository.findById(warehouseReceiptDetailEntity.getId()).isPresent()) {
+                warehouseReceiptDetailService.updateWarehouseReceiptDetail(warehouseReceiptDetailEntity);
+                response = ResponseUtils.buildMessageReponse(HttpStatus.OK, "Update Successfully");
+            } else {
+                warehouseReceiptDetailService.updateWarehouseReceiptDetail(warehouseReceiptDetailEntity);
+                response = ResponseUtils.buildMessageReponse(HttpStatus.CREATED, "Successfully Added");
+            }
 
         } else {
             response = ResponseUtils.buildMessageReponse(HttpStatus.BAD_REQUEST, message);
@@ -58,10 +65,17 @@ public class WarehouseReceiptDetailController {
         return response;
     }
 
-    @DeleteMapping("/warehousereceiptdetail/delete/{id}")
+    @DeleteMapping("/{id}")
     public ResponseEntity<?> deleteWarehouseReceiptDetail(@PathVariable("id") Integer id) {
-        warehouseReceiptDetailRepository.deleteById(id);
-        return new ResponseEntity<>(HttpStatus.OK);
+        ResponseEntity<Object> response = ResponseEntity.status(HttpStatus.OK).build();
+        if (warehouseReceiptDetailRepository.findById(id).isPresent()) {
+            warehouseReceiptDetailRepository.deleteById(id);
+            response = ResponseUtils.buildMessageReponse(HttpStatus.OK, "Successfully Deleted");
+        } else {
+            response = ResponseUtils.buildMessageReponse(HttpStatus.NOT_FOUND, "WareHouseReceiptDetail was not found");
+        }
+        return response;
+
     }
 
     private String inputcheck(WarehouseReceiptDetailDto warehouseReceiptDetailDto) {
@@ -71,17 +85,20 @@ public class WarehouseReceiptDetailController {
         }
 
         StringBuilder errorField = new StringBuilder();
-        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getWarehousereceiptid())) {
+        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getId())) {
+            errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "Id" : ",Id"));
+        }
+        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getWarehouse_receipt_id())) {
             errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "WarehouseReceiptId" : ",WarehouseReceiptId"));
         }
-        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getPizzadetailid())) {
+        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getPizza_detail_id())) {
             errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "PizzaDetailId" : ",PizzaDetailId"));
         }
         if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getPrice())) {
             errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "Price" : ",Price"));
         }
-        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getAmount()) && ValidateUtils.isFullWidthDigit(String.valueOf(warehouseReceiptDetailDto.getAmount()))) {
-            errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "Amount" : ",Amount"));
+        if (ValidateUtils.isNullOrEmpty(warehouseReceiptDetailDto.getQuantity()) && ValidateUtils.isFullWidthDigit(String.valueOf(warehouseReceiptDetailDto.getQuantity()))) {
+            errorField.append((ValidateUtils.isNullOrEmpty(errorField) ? "Quantity" : ",Quantity"));
         }
         String message = "";
         if (!ValidateUtils.isNullOrEmpty((errorField.toString()))) {
